DAN-361 Fixed tmobile project view callback
